1 курс/Введение в профессиональную деятельность/server/requirements.txt
requests
jsonschema
pyjwt
werkzeug
pymongo
flask
dnspython
gunicorn
flask_cors
nanoid
pytest
pytest-flask